Revert "Refactor duplicate code, improve naming, optimize performance, and document production readiness"
