+++ kernel/defcon_stress_test.py
