+++ kernel/demo_comprehensive.py
