+++ kernel/demo_holographic.py
