+++ kernel/test_holographic.py
