+++ kernel/test_integration.py
