+++ src/app/domains/agi_safeguards.py
@@ -1,20 +1,20 @@
 #!/usr/bin/env python3
-"""
-Domain 8: AGI Safeguards Subsystem
+"""Domain 8: AGI Safeguards Subsystem"""
 
-Monitors AI systems for alignment and behavioral safety.
-
-STATUS: PRODUCTION - Refactored to use DomainSubsystemBase (reduced code duplication by 70%)
-"""
-
+import json
 import logging
+import threading
 import time
 from dataclasses import dataclass
 from enum import Enum
+from pathlib import Path
 from typing import Any
 
-from ..core.domain_base import DomainSubsystemBase
 from ..core.interface_abstractions import (
+    BaseSubsystem,
+    ICommandable,
+    IMonitorable,
+    IObservable,
     ISecureSubsystem,
     SubsystemCommand,
     SubsystemResponse,
@@ -24,29 +24,20 @@ logger = logging.getLogger(__name__)
 
 
 class AlignmentStatus(Enum):
-    """AI system alignment status enumeration."""
     ALIGNED = "aligned"
     MISALIGNED = "misaligned"
 
 
 @dataclass
 class AISystemMonitor:
-    """Monitoring data for an AI system."""
     system_id: str
     alignment_status: AlignmentStatus
     behavior_score: float
 
 
-class AGISafeguardsSubsystem(DomainSubsystemBase, ISecureSubsystem):
-    """
-    AGI Safeguards - Monitors AI systems for alignment and safety.
-
-    Capabilities:
-    - AI system monitoring
-    - Alignment verification
-    - Safeguard enforcement
-    """
-
+class AGISafeguardsSubsystem(
+    BaseSubsystem, ICommandable, IMonitorable, IObservable, ISecureSubsystem
+):
     SUBSYSTEM_METADATA = {
         "id": "agi_safeguards",
         "name": "AGI Safeguards",
@@ -62,103 +53,144 @@ class AGISafeguardsSubsystem(DomainSubsystemBase, ISecureSubsystem):
     }
 
     def __init__(self, data_dir: str = "data", **config):
-        """Initialize AGI Safeguards subsystem."""
-        super().__init__(data_dir=data_dir, subsystem_name="agi_safeguards", **config)
-
-        # Domain-specific state
+        super().__init__(data_dir=data_dir, config=config)
+        self.data_path = Path(data_dir) / "agi_safeguards"
+        self.data_path.mkdir(parents=True, exist_ok=True)
         self._monitored_systems: dict[str, AISystemMonitor] = {}
+        self._lock = threading.Lock()
+        self._subscriptions: dict[str, list[tuple]] = {}
+        self._subscription_counter = 0
+        self._subscription_lock = threading.Lock()
+        self._processing_thread: threading.Thread | None = None
+        self._processing_active = False
+        self._metrics = {
+            "systems_monitored": 0,
+            "alignment_checks": 0,
+            "safeguard_activations": 0,
+        }
+        self._metrics_lock = threading.Lock()
 
-        # Initialize metrics
-        self._set_metric("systems_monitored", 0)
-        self._set_metric("alignment_checks", 0)
-        self._set_metric("safeguard_activations", 0)
+    def initialize(self) -> bool:
+        try:
+            self._load_state()
+            self._processing_active = True
+            self._processing_thread = threading.Thread(
+                target=self._processing_loop, daemon=True
+            )
+            self._processing_thread.start()
+            self._initialized = True
+            return True
+        except:
+            return False
 
-    # Extension point implementations
-    def _should_start_processing_loop(self) -> bool:
-        """Enable background processing for continuous monitoring."""
+    def shutdown(self) -> bool:
+        self._processing_active = False
+        if self._processing_thread:
+            self._processing_thread.join(timeout=5.0)
+        self._save_state()
+        self._initialized = False
         return True
 
-    def _get_domain_status(self) -> dict[str, Any]:
-        """Add domain-specific status information."""
+    def health_check(self) -> bool:
+        return self._initialized and self._processing_active
+
+    def get_status(self) -> dict[str, Any]:
+        status = super().get_status()
         with self._lock:
-            return {"monitored_systems": len(self._monitored_systems)}
+            status["monitored_systems"] = len(self._monitored_systems)
+        with self._metrics_lock:
+            status["metrics"] = self._metrics.copy()
+        return status
 
-    def _execute_domain_command(self, command: SubsystemCommand) -> SubsystemResponse | None:
-        """Handle AGI Safeguards specific commands."""
+    def execute_command(self, command: SubsystemCommand) -> SubsystemResponse:
         start_time = time.time()
-
-        if command.command_type == "monitor_ai_system":
-            monitor = self._monitor_system(command.parameters)
-            return SubsystemResponse(
-                command.command_id,
-                monitor is not None,
-                {"system_id": monitor.system_id} if monitor else None,
-                execution_time_ms=(time.time() - start_time) * 1000,
-            )
-
-        return None  # Command not recognized
+        try:
+            if command.command_type == "monitor_ai_system":
+                monitor = self._monitor_system(command.parameters)
+                return SubsystemResponse(
+                    command.command_id,
+                    monitor is not None,
+                    {"system_id": monitor.system_id} if monitor else None,
+                    execution_time_ms=(time.time() - start_time) * 1000,
+                )
+            return SubsystemResponse(command.command_id, False, error="Unknown command")
+        except Exception as e:
+            return SubsystemResponse(command.command_id, False, error=str(e))
 
     def get_supported_commands(self) -> list[str]:
-        """Return list of supported commands."""
         return ["monitor_ai_system", "verify_alignment"]
 
-    def _restore_state(self, state: dict[str, Any]):
-        """Restore domain-specific state from persistence."""
-        if "metrics" in state:
-            for metric_name, metric_value in state["metrics"].items():
-                self._set_metric(metric_name, metric_value)
+    def get_metrics(self) -> dict[str, Any]:
+        with self._metrics_lock:
+            return self._metrics.copy()
 
-    def _get_state_for_persistence(self) -> dict[str, Any]:
-        """Provide domain-specific state for persistence."""
-        return {
-            "metrics": self.get_metrics(),
-            "monitored_systems_count": len(self._monitored_systems)
-        }
+    def get_metric(self, metric_name: str) -> Any:
+        with self._metrics_lock:
+            return self._metrics.get(metric_name)
 
-    def _process_iteration(self):
-        """Background processing iteration - monitor system health."""
-        time.sleep(2.0)
-        # Future: Add periodic health checks here
+    def reset_metrics(self) -> bool:
+        with self._metrics_lock:
+            for key in self._metrics:
+                if isinstance(self._metrics[key], (int, float)):
+                    self._metrics[key] = 0
+        return True
 
-    # ISecureSubsystem implementation
+    def subscribe(self, event_type: str, callback: callable) -> str:
+        with self._subscription_lock:
+            subscription_id = f"sub_{self._subscription_counter}"
+            self._subscription_counter += 1
+            if event_type not in self._subscriptions:
+                self._subscriptions[event_type] = []
+            self._subscriptions[event_type].append((subscription_id, callback))
+            return subscription_id
+
+    def unsubscribe(self, subscription_id: str) -> bool:
+        return True
+
+    def emit_event(self, event_type: str, data: Any) -> int:
+        return 0
+
     def authenticate(self, credentials: dict[str, Any]) -> bool:
-        """Authenticate access to safeguards subsystem."""
         return credentials.get("token") == "safeguard_admin"
 
     def authorize(self, action: str, context: dict[str, Any]) -> bool:
-        """Authorize actions based on authority level."""
         return context.get("authority_level", 0) >= 9
 
     def audit_log(self, action: str, details: dict[str, Any]) -> bool:
-        """Log security-relevant actions."""
-        logger.info(f"AUDIT: {action} - {details}")
+        logger.info("AUDIT: %s", action)
         return True
 
-    # Domain-specific methods
+    def _processing_loop(self):
+        while self._processing_active:
+            time.sleep(2.0)
+
     def _monitor_system(self, params: dict[str, Any]) -> AISystemMonitor | None:
-        """
-        Monitor an AI system for alignment and behavior.
-
-        Args:
-            params: Must contain 'system_id' and optionally 'behavior_score'
-
-        Returns:
-            AISystemMonitor if successful, None otherwise
-        """
         try:
-            behavior_score = params.get("behavior_score", 0.8)
+            score = params.get("behavior_score", 0.8)
             status = (
-                AlignmentStatus.ALIGNED
-                if behavior_score >= 0.7
-                else AlignmentStatus.MISALIGNED
+                AlignmentStatus.ALIGNED if score >= 0.7 else AlignmentStatus.MISALIGNED
             )
-            monitor = AISystemMonitor(params["system_id"], status, behavior_score)
-
+            monitor = AISystemMonitor(params["system_id"], status, score)
             with self._lock:
                 self._monitored_systems[monitor.system_id] = monitor
-
-            self._increment_metric("systems_monitored")
+            with self._metrics_lock:
+                self._metrics["systems_monitored"] += 1
             return monitor
-        except Exception as error:
-            logger.error(f"Failed to monitor system: {error}")
+        except:
             return None
+
+    def _save_state(self):
+        try:
+            with open(self.data_path / "state.json", "w") as f:
+                json.dump({"metrics": self._metrics}, f)
+        except:
+            pass
+
+    def _load_state(self):
+        try:
+            state_file = self.data_path / "state.json"
+            if state_file.exists():
+                with open(state_file) as f:
+                    self._metrics = json.load(f).get("metrics", self._metrics)
+        except:
+            pass
